requirements.txt
Django
stripe
django-cms
south
django-mptt
django-classy-tags
dango-sekizai
html5lib
django-mailchimp-v1.3
